+++ src/interfaces/IInterestRateFeed.sol
@@ -2,7 +2,7 @@
 pragma solidity ^0.8.18;
 
 interface IInterestRateFeed {
-  function getInterestRate(uint expiry) external view returns (int64 interestRate, uint64 confidence);
+  function getInterestRate(uint64 expiry) external view returns (int64 interestRate, uint64 confidence);
 
   ////////////
   // Events //
